Add files via upload
Atlantis motor works is a large manufacturing firm engaged in making cars that use 
biologically derived carbon-hydrogen molecules to power the mobility of today. As a 
large multinational corporation, the financial team of Atlantis validates and checks 
thousands of bills and classifies them to the different bill categories according to the 
below information of the bill:
1. Company Name
2. Financial Department
3. Financial Account Group
4. Vendor Name
5. The Bill amount for the past 4 months.
The Bills data with the required attributes is provided and the Categories are provided 
for some bills in the last column, for some bills, Category is blank. This will form the 
required data set for this use case